+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+0.29 (2024-09-25)
+===============
+* [@BartoszDziurowicz](https://github.com/BartekDziurowicz/carbon-tracker-web/pull/19)
+* Create reports ctx
+* Create logic to handle new indicator endpoint
+* Develop reports component with select, label and button sub-components
+* Implement new api call to retrieve calculated indicators
+* Extend reports context by calculatedIndicators and setCalculatedIndicators
+* Implement logic for generate button
+
 0.28 (2024-09-17)
 ===============
 * [@BartoszDziurowicz](https://github.com/BartekDziurowicz/carbon-tracker-web/pull/18)